# backend/api-gateway-gcp/openapi2-run.yaml
...
swagger: "2.0"
info:
  title: <API_ID>
  description: API Gateway routes to project Cloud Run backend
  version: 1.0.0
schemes:
  - http
produces:
  - application/json

# FIREBASE AUTH
securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://securetoken.google.com/<PROJECT_ID>"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "<PROJECT_ID>"
paths:
...
  /files:
    options:
      summary: CORS pre-flight
      operationId: corsFiles
      x-google-backend:
        address: <MEDIA_SERVICE_URL>/
      responses:
        "200":
          description: A successful response
    get:
      summary: Fetch All Files Uploads
      x-google-backend:
        address: <MEDIA_SERVICE_URL>/api/files
      operationId: fetchAllFiles
      security:
        - firebase: []
      responses:
        "200":
          description: Successfully fetched all file upload
          schema:
            type: array
            items:
              type: object
              properties:
                name:
                  type: string
                url:
                  type: string
                type:
                  type: string
        "500":
          description: Cloud storage access failed
          schema:
            type: string
    post:
      summary: Upload Files
      x-google-backend:
        address: <MEDIA_SERVICE_URL>/api/files
      operationId: uploadFiles
      security:
        - firebase: []
      # https://swagger.io/docs/specification/2-0/file-upload/
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: formData
          type: file
          description: The file to upload
      responses:
        "200":
          description: File upload success
          schema:
            type: string
        "500":
          description: Cloud storage access failed
          schema:
            type: string

  # PROFILE SERVICE ROUTES
...